Clean up 4242/4244/4267 warnings in parser and byteCodeGen
     This changeset cleans up conversion warnings in parser, bytecode generator (including AsmJs), and in regex compiler.
     The fixes fall in one of the categories below:
     - add explicit conversion and guard them with assertions or checks as appropriate
     - change data types to be consistent with producers/consumers so warnings go away

reviewed by: hiteshk, goyakin, michhol, curtism

+++ lib/Parser/ptree.h
@@ -227,7 +227,7 @@ struct PnFnc
     uint nestedCount; // Nested function count (valid until children have been processed)
     uint nestedIndex; // Index within the parent function
 
-    uint firstDefaultArg; // Position of the first default argument, if any
+    uint16 firstDefaultArg; // Position of the first default argument, if any
 
     unsigned int fncFlags;
     long astSize;
